RoogleDocs

Programmatically substitute images, data and tables into a google doc.

Version: 0.2.1

Generated: 2023-09-29T16:53:46.640578179

Arguments

tokenDirectory

tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to `.tokenDirectory()`) - (java expects a String)

disabled

disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to `getOption('roogledocs.disabled',FALSE)`) - (java expects a boolean)

Details

R library to perform limited interactions with google docs (and maybe one day slides) in R via the Java API library. The purpose being to support google docs as a platform for interactive development and documentation of data analysis in R for scientific publication, although it is not limited to this purpose. The workflow supported is a parallel documentation and analysis where a team of people are working collaboratively on documentation, whilst at the same time analysis is being performed and results updated repeatedly as a result of new data. In this environment updating numeric results, tabular data and figures in word documents manually becomes annoying. With roogledocs you can automate this a bit like a RMarkdown document, but with the added benefit that the content can be updated independently of the analysis, by the wider team.

Methods

Constructors

J$RoogleDocs$new(tokenDirectory, disabled)

Class methods

instance$enable()

instance$disable()

instance$withDocument(shareUrlOrDocId)

instance$findOrCreateDocument(title)

instance$findOrCloneTemplate(title, templateUri)

instance$tagsDefined()

instance$updateTaggedText(text, tagName)

instance$updateTaggedImage(absoluteFilePath, tagName, dpi)

instance$revertTags()

instance$removeTags(confirm)

instance$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)

instance$updateFigure(absoluteFilePath, figureIndex, dpi)

instance$saveAsPdf(absoluteFilePath, uploadCopy)

instance$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)

instance$appendText(text, style)

instance$appendFormattedParagraph(formattedTextDf)

instance$updateCitations(bibTex, citationStyle)

instance$clone()

instance$print()

Constructor new()

Create a Roogledocs object for managing the interaction.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
        

Arguments

tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to `.tokenDirectory()`) - (java expects a String)

disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to `getOption('roogledocs.disabled',FALSE)`) - (java expects a boolean)

Returns

the new R6 RoogleDocs object

Method enable()

Enables roogledocs method calls for this document.

It is likely one of `withDocument()`, `findOrCreateDocument()` or `findOrCloneTemplate()` methods will be needed to specify the document.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$enable()
        

Arguments

none

Returns

R6 RoogleDocs object: itself - a fluent method

Method disable()

Disables roogledocs temporarily for this document.

While disabled all calls to roogledocs will silently abort.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$disable()
        

Arguments

none

Returns

R6 RoogleDocs object: itself - a fluent method

Method withDocument()

Select a document by its share url or id.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$withDocument(shareUrlOrDocId)
        

Arguments

shareUrlOrDocId the url from clicking a share button in google docs or an id from searchForDocuments() method - (java expects a String)

Returns

R6 RoogleDocs object: itself - a fluent method

Method findOrCreateDocument()

Search for a document by name or create one if missing.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$findOrCreateDocument(title)
        

Arguments

title a document title. If there is an exact match in google drive then that document will be used otherwise a new one will be created. - (java expects a String)

Returns

R6 RoogleDocs object: itself - a fluent method

Method findOrCloneTemplate()

Get a document by name or create one from a template if missing.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$findOrCloneTemplate(title, templateUri)
        

Arguments

title a document title. If there is an exact match in google drive then that document will be used otherwise a new one will be created. - (java expects a String)

templateUri the share link (or document id) of a template google document - (java expects a String)

Returns

R6 RoogleDocs object: itself - a fluent method

Method tagsDefined()

List all tags

Finds tags defined in the current document

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$tagsDefined()
        

Arguments

none

Returns

RDataframe: a dataframe containing tag and count columns

Method updateTaggedText()

Relace tags for text

Substitutes all occurrences of tag-name with the text parameter.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$updateTaggedText(text, tagName)
        

Arguments

text the value to replace the tag with (e.g. a result from analysis) (cannot be empty) - (java expects a String)

tagName the tag name - (defaulting to `deparse(substitute(text))`) - (java expects a String)

Returns

R6 RoogleDocs object: itself - a fluent method

Method updateTaggedImage()

Replace a tag with an image.

Substitutes all occurrences of tag-name with an image from the local storage. There are limited circumstances in which using this is a good idea. It will almost always be better to use `updateFigure()` to insert an image by index. If you choose to ignore this warning, beware combining this with `updateFigure()` as potentially the figure indexes may change dynamically.

The image is uploaded to your google drive as a temporary file, and briefly made publically readable. From there it is inserted into the google doc, and one completed the temporary file deleted from your google drive.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$updateTaggedImage(absoluteFilePath, tagName, dpi)
        

Arguments

absoluteFilePath a file path to an png image file. - (java expects a String)

tagName the tag name - (defaulting to `deparse(substitute(absoluteFilePath))`) - (java expects a String)

dpi the dots per inch of the image in the document (defaults to 300) - (defaulting to `300`) - (java expects a double)

Returns

R6 RoogleDocs object: itself - a fluent method

Method revertTags()

Revert tagged text and images.

Remove all tagged text and images inserted by roogledocs and returns the bare document the tags in place. This does not affect figures and tables inserted by index (i.e. without tags) This is needed if content is being moved around as cut and paste of tagged content unfortunately removes the internal named range of the tag.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$revertTags()
        

Arguments

none

Returns

R6 RoogleDocs object: itself - a fluent method

Method removeTags()

Remove all tags

Finds tags defined in the current document and removes them. This cannot be undone.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$removeTags(confirm)
        

Arguments

confirm - This action must be confirmed by passing `true` as cannot be undone. - (defaulting to `(menu(c('Yes','No'), title = 'Are you...`) - (java expects a boolean)

Returns

R6 RoogleDocs object: itself - a fluent method

Method updateTable()

Update or insert a formatted table into the document.

The table and formatting are described in a dataframe the format of which is documented in the as.long_format_table() method.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)
        

Arguments

longFormatTable A dataframe consisting of the table content and formatting indexed by row and column. at a minimum this should have columns label,row,col, but may also include rowSpan,colSpan,fillColour, leftBorderWeight, rightBorderWeight, topBorderWeight, bottomBorderWeight, alignment (START,CENTER,END), valignment (TOP,MIDDLE,BOTTOM), fontName, fontFace, fontSize. - (java expects a RDataframe)

tableIndex what is the table index in the document? This can be left out for a new table at the end of the document. - (defaulting to `-1`) - (java expects a int)

colWidths A vector including the relative length of each column. This can be left out if longFormatTable comes from as.long_format_table - (defaulting to `attr(longFormatTable,'colWidths')`) - (java expects a RNumericVector)

tableWidthInches The final width of the table in inches (defaults to a size that fits in A4 page with margins) - (defaulting to `6.2`) - (java expects a RNumeric)

Returns

R6 RoogleDocs object: itself - a fluent method

Method updateFigure()

Update or insert a figure in the document from a locally stored PNG.

This function uploads the image into a temporary file onto your Google Drive, and makes it briefly publically readable. From there inserts it into the google document. Once this is complete the temporary google drive copy of the image is deleted.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$updateFigure(absoluteFilePath, figureIndex, dpi)
        

Arguments

absoluteFilePath a file path to an png image file (only png is supported at this point). - (java expects a String)

figureIndex what is the figure index in the document? (This only counts inline images - and ignores absolutely positioned ones). leave out for a new image at the end of the document. - (defaulting to `-1`) - (java expects a int)

dpi the dots per inch of the image in the document (defaults to 300). the final size of the image in the doc will be determined by the image file dimensions and the dpi. - (defaulting to `300`) - (java expects a double)

Returns

R6 RoogleDocs object: itself - a fluent method

Method saveAsPdf()

Save the document as a PDF

Saves a snapshot of the current google doc with `roogledocs` links removed as a pdf to a local drive. This is mainly intended for snapshotting the current state of the document. For final export once all analysis is complete it may be preferable to call `doc$removeTags()` and manually export the output

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$saveAsPdf(absoluteFilePath, uploadCopy)
        

Arguments

absoluteFilePath - a file path to save the pdf. - (java expects a String)

uploadCopy place a copy of the downloaded pdf back onto google drive in the same folder as the document for example for keeping submitted versions of a updated document. This will overwrite files of the same name in the google drive directory. - (defaulting to `FALSE`) - (java expects a boolean)

Returns

R6 RoogleDocs object: itself - a fluent method

Method uploadSupplementaryFiles()

Upload a file into the same directory as the document.

This allow you to load e.g. a supplementary file, or the pdf of an image file or a docx/html version of a table into google drive into the same directory as the document you are editing. This is handy for organising all the files for a journal submission in one place. Any kind of file can be loaded, and the mimetype will be detected. Normal Google Drive rules for uploads will be triggered at this point. As google drive can have multiple files with the same name the behaviour if the file already exists is slightly complex, with `overwrite` and `duplicate` options.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)
        

Arguments

absoluteFilePath - a file path to upload. - (java expects a String)

overwrite - if matching file(s) are found in the target, delete them before uploading the new one. - (defaulting to `FALSE`) - (java expects a boolean)

duplicate - if matching file(s) are found in the target, upload this new file anyway, creating duplicate names in the folder. - (defaulting to `FALSE`) - (java expects a boolean)

Returns

R6 RoogleDocs object: itself - a fluent method

Method appendText()

Append text to the document with optional paragraph styling. If you run text blocks into each other without newlines the whole resulting paragraph will be styled. You would normally not want this so it is up to you to end paragraphs with a new line character, before changing styles.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$appendText(text, style)
        

Arguments

text - a single string with the text to append which may include newlines - (java expects a RCharacter)

style - one of NORMAL_TEXT, TITLE, SUBTITLE, HEADING_1, ... HEADING_6 - (defaulting to `'NORMAL_TEXT'`) - (java expects a RCharacter)

Returns

R6 RoogleDocs object: itself - a fluent method

Method appendFormattedParagraph()

Append a new paragraph, with text from the 'label' column with optional formating in the other columns.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$appendFormattedParagraph(formattedTextDf)
        

Arguments

formattedTextDf - a data frame containing the columns label, and optionally: link (as a URL), fontName, fontFace, fontSize. - (java expects a RDataframe)

Returns

R6 RoogleDocs object: itself - a fluent method

Method updateCitations()

Update citation tags in the document.

A citation tag is like this cite:challen2020;danon2021. The ids

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(tokenDirectory, disabled);
instance$updateCitations(bibTex, citationStyle)
        

Arguments

bibTex - a string containing the bibtex - (java expects a RCharacter)

citationStyle - the CSL specification - (defaulting to `'ieee'`) - (java expects a RCharacter)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

## -----------------------------------
## Construct new instance of RoogleDocs
## -----------------------------------
if (FALSE) {
J = roogledocs::JavaApi$get()
# appropriate parameter values must be provided
instance = J$RoogleDocs$new(tokenDirectory, disabled)
}

## -----------------------------------
## Method `RoogleDocs$enable()`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$enable()
}

## -----------------------------------
## Method `RoogleDocs$disable()`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$disable()
}

## -----------------------------------
## Method `RoogleDocs$withDocument(shareUrlOrDocId)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$withDocument(shareUrlOrDocId)
}

## -----------------------------------
## Method `RoogleDocs$findOrCreateDocument(title)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$findOrCreateDocument(title)
}

## -----------------------------------
## Method `RoogleDocs$findOrCloneTemplate(title, templateUri)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$findOrCloneTemplate(title, templateUri)
}

## -----------------------------------
## Method `RoogleDocs$tagsDefined()`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$tagsDefined()
}

## -----------------------------------
## Method `RoogleDocs$updateTaggedText(text, tagName)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$updateTaggedText(text, tagName)
}

## -----------------------------------
## Method `RoogleDocs$updateTaggedImage(absoluteFilePath, tagName, dpi)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$updateTaggedImage(absoluteFilePath, tagName, dpi)
}

## -----------------------------------
## Method `RoogleDocs$revertTags()`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$revertTags()
}

## -----------------------------------
## Method `RoogleDocs$removeTags(confirm)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$removeTags(confirm)
}

## -----------------------------------
## Method `RoogleDocs$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)
}

## -----------------------------------
## Method `RoogleDocs$updateFigure(absoluteFilePath, figureIndex, dpi)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$updateFigure(absoluteFilePath, figureIndex, dpi)
}

## -----------------------------------
## Method `RoogleDocs$saveAsPdf(absoluteFilePath, uploadCopy)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$saveAsPdf(absoluteFilePath, uploadCopy)
}

## -----------------------------------
## Method `RoogleDocs$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)
}

## -----------------------------------
## Method `RoogleDocs$appendText(text, style)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$appendText(text, style)
}

## -----------------------------------
## Method `RoogleDocs$appendFormattedParagraph(formattedTextDf)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$appendFormattedParagraph(formattedTextDf)
}

## -----------------------------------
## Method `RoogleDocs$updateCitations(bibTex, citationStyle)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$updateCitations(bibTex, citationStyle)
}