RoogleSlides

Programmatically substitute images, data and tables into a google presentation.

Version: 0.3.0

Generated: 2024-01-17T13:26:56.206521975

Arguments

tokenDirectory

tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to `.tokenDirectory()`) - (java expects a String)

disabled

disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to `getOption('roogledocs.disabled',FALSE)`) - (java expects a boolean)

Details

The purpose being to support google slides as a platform for interactive development and documentation of data analysis in R. The workflow supported is a parallel documentation and analysis where a team of people are working collaboratively on documentation, whilst at the same time analysis is being performed and results updated repeatedly as a result of new data. In this environment updating numeric results, tabular data and figures in word documents manually becomes annoying. With roogledocs you can automate this a bit like a RMarkdown document, but with the added benefit that the content can be updated independently of the analysis, by the wider team.

Methods

Constructors

J$RoogleSlides$new(tokenDirectory, disabled)

Class methods

instance$enable()

instance$disable()

instance$withDocument(shareUrlOrDocId)

instance$findOrCreateDocument(title)

instance$findOrCloneTemplate(title, templateUri)

instance$tagsDefined()

instance$updateTaggedText(text, tagName)

instance$updateTaggedImage(absoluteFilePath, tagName, keepUpload)

instance$updateTaggedTable(longFormatTable, tagName, colWidths)

instance$removeTags(confirm)

instance$saveAsPdf(absoluteFilePath, uploadCopy)

instance$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)

instance$appendFormattedSlide(title, formattedTextDf, layout)

instance$slideDimensions()

instance$slideLayouts()

instance$updateCitations(bibTexPath, citationStyle)

instance$clone()

instance$print()

Constructor new()

Create a RoogleSlides object for managing the interaction.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
        

Arguments

tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to `.tokenDirectory()`) - (java expects a String)

disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to `getOption('roogledocs.disabled',FALSE)`) - (java expects a boolean)

Returns

the new R6 RoogleSlides object

Method enable()

Enables roogledocs method calls for this document.

It is likely one of `withDocument()`, `findOrCreateDocument()` or `findOrCloneTemplate()` methods will be needed to specify the document.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$enable()
        

Arguments

none

Returns

R6 RoogleSlides object: itself - a fluent method

Method disable()

Disables roogledocs temporarily for this document.

While disabled all calls to roogledocs will silently abort.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$disable()
        

Arguments

none

Returns

R6 RoogleSlides object: itself - a fluent method

Method withDocument()

Select a document by its share url or id.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$withDocument(shareUrlOrDocId)
        

Arguments

shareUrlOrDocId the url from clicking a share button in google slides or an id from searchForDocuments() method - (java expects a String)

Returns

R6 RoogleSlides object: itself - a fluent method

Method findOrCreateDocument()

Search for a document by name or create one if missing.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$findOrCreateDocument(title)
        

Arguments

title a document title. If there is an exact match in google drive then that document will be used otherwise a new one will be created. - (java expects a String)

Returns

R6 RoogleSlides object: itself - a fluent method

Method findOrCloneTemplate()

Get a document by name or create one from a template if missing.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$findOrCloneTemplate(title, templateUri)
        

Arguments

title a document title. If there is an exact match in google drive then that document will be used otherwise a new one will be created. - (java expects a String)

templateUri the share link (or document id) of a template google document - (java expects a String)

Returns

R6 RoogleSlides object: itself - a fluent method

Method tagsDefined()

List all tags

Finds tags defined in the current document

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$tagsDefined()
        

Arguments

none

Returns

RDataframe: a dataframe containing tag and count columns

Method updateTaggedText()

Replace tags for text

Substitutes all occurrences of tag-name with the text parameter. If the tag is not found then a new slide is inserted at the end in a section titled "Unmatched tags:". From there they can be cut and pasted into the right place.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$updateTaggedText(text, tagName)
        

Arguments

text the value to replace the tag with (e.g. a result from analysis) (cannot be empty) - (java expects a String)

tagName the tag name - (defaulting to `deparse(substitute(text))`) - (java expects a String)

Returns

R6 RoogleSlides object: itself - a fluent method

Method updateTaggedImage()

Replace a tag with an image.

Substitutes all occurrences of tag-name with an image from the local storage.

The image is uploaded to your google drive as a temporary file, and made publicly readable. From there it is inserted into the google slides, and once completed the temporary file deleted from your google drive, unless `keepUpload` is true. Insertion is done in the dimensions of the containing box of the image if it already exists or a default slide body box if not.

If the tag is not found in the document a new slide will be created at the end of the presentation with the image and an uninformative title which can be changed.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$updateTaggedImage(absoluteFilePath, tagName, keepUpload)
        

Arguments

absoluteFilePath a file path to an png image file. - (java expects a String)

tagName the tag name - (defaulting to `deparse(substitute(absoluteFilePath))`) - (java expects a String)

keepUpload keep the uploaded image as a supplementary file in the same directory as the google doc. N.B. the result will be publicly readable. - (defaulting to `FALSE`) - (java expects a boolean)

Returns

R6 RoogleSlides object: itself - a fluent method

Method updateTaggedTable()

Replace a tag with a table.

Substitutes a unique occurrence of tag-name with a table. The tag must either be in a text box shape or as the first entry in a table. Once inserted the table is tagged using a zero width character as the very first item in the first cell. This will be removed if `removeTags()` is called.

If the tag is not found in the document a new slide will be created at the end with the table.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$updateTaggedTable(longFormatTable, tagName, colWidths)
        

Arguments

longFormatTable A dataframe consisting of the table content and formatting indexed by row and column. at a minimum this should have columns label,row,col, but may also include rowSpan,colSpan,fillColour, leftBorderWeight, rightBorderWeight, topBorderWeight, bottomBorderWeight, alignment (START,CENTER,END), valignment (TOP,MIDDLE,BOTTOM), fontName, fontFace, fontSize. - (java expects a RDataframe)

tagName the tag name - (defaulting to `deparse(substitute(longFormatTable))`) - (java expects a String)

colWidths A vector including the relative length of each column. This can be left out if longFormatTable comes from as.long_format_table - (defaulting to `attr(longFormatTable,'colWidths')`) - (java expects a RNumericVector)

Returns

R6 RoogleSlides object: itself - a fluent method

Method removeTags()

Remove all tags

Finds tags defined in the current document and removes them. This cannot be undone, except by rolling back to a previous version.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$removeTags(confirm)
        

Arguments

confirm - This action must be confirmed by passing `true` as cannot be undone. - (defaulting to `(menu(c('Yes','No'), title = 'Are you...`) - (java expects a boolean)

Returns

R6 RoogleSlides object: itself - a fluent method

Method saveAsPdf()

Save the document as a PDF

Saves a snapshot of the current google slides with `roogledocs` links removed as a pdf to a local drive. This is mainly intended for snap-shotting the current state of the document. For final export once all analysis is complete it may be preferable to call `doc$removeTags()` and manually export the output but after this no further updating is possible.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$saveAsPdf(absoluteFilePath, uploadCopy)
        

Arguments

absoluteFilePath - a file path to save the pdf. - (java expects a String)

uploadCopy place a copy of the downloaded pdf back onto google drive in the same folder as the document for example for keeping submitted versions of a updated document. This will overwrite files of the same name in the google drive directory. - (defaulting to `FALSE`) - (java expects a boolean)

Returns

R6 RoogleSlides object: itself - a fluent method

Method uploadSupplementaryFiles()

Upload a file into the same directory as the document.

This allow you to load e.g. a supplementary file, or the pdf of an image file or a docx/html version of a table into google drive into the same directory as the slides you are editing. This is handy for organising all the files for a journal submission in one place. Any kind of file can be loaded, and the mimetype will be detected. Normal Google Drive rules for uploads will be triggered at this point. As google drive can have multiple files with the same name the behaviour if the file already exists is slightly complex, with `overwrite` and `duplicate` options.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)
        

Arguments

absoluteFilePath - a file path to upload. - (java expects a String)

overwrite - if matching file(s) are found in the target, delete them before uploading the new one. - (defaulting to `FALSE`) - (java expects a boolean)

duplicate - if matching file(s) are found in the target, upload this new file anyway, creating duplicate names in the folder. - (defaulting to `FALSE`) - (java expects a boolean)

Returns

R6 RoogleSlides object: itself - a fluent method

Method appendFormattedSlide()

Append a new "TITLE_AND_BODY" slide, with formatted text from the 'label' column with optional formating in the other columns.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$appendFormattedSlide(title, formattedTextDf, layout)
        

Arguments

title - A plain text title - (java expects a String)

formattedTextDf - a data frame containing the columns label, and optionally: link (as a URL), fontName, fontFace, fontSize. - (java expects a RDataframe)

layout - (defaulting to `'TITLE_AND_BODY'`) - (java expects a String)

Returns

R6 RoogleSlides object: itself - a fluent method

Method slideDimensions()

The default dimensions of the body of a new "TITLE_AND_BODY" slide.

This is defined as the second text box of the default "new slide" and this is the place that images or tables will be placed by default. If you want to fit in with the theme then images will need to be to be sized to these dimensions to fill the slide.

This will throw errors if your slides do not define a "TITLE_AND_BODY" layout and may behave unpredictably if the content is not in the second box as defined in the template. In other words this is a guide.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$slideDimensions()
        

Arguments

none

Returns

RNamedList: a named list with width and height in inches

Method slideLayouts()

The layouts available in the slides templates

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$slideLayouts()
        

Arguments

none

Returns

RCharacterVector: a list of available slide layouts

Method updateCitations()

Update citation tags in the document.

A citation tag is like this `cite:challen2020;danon2021`. The ids are matched against the provided bibtex, and the tags are replaced with an appropriate citation string. The bibliography itself is added to a specific slide for references which can be decided with the `references` tag.

If references do not already exist and there if no `references` tag a new slide will be created at the end of the presentation.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleSlides$new(tokenDirectory, disabled);
instance$updateCitations(bibTexPath, citationStyle)
        

Arguments

bibTexPath - the full file path to the file containing the bibtex - (java expects a RCharacter)

citationStyle - the CSL specification - (defaulting to `'ieee'`) - (java expects a RCharacter)

Returns

R6 RoogleSlides object: itself - a fluent method

Examples

## -----------------------------------
## Construct new instance of RoogleSlides
## -----------------------------------
if (FALSE) {
J = roogledocs::JavaApi$get()
# appropriate parameter values must be provided
instance = J$RoogleSlides$new(tokenDirectory, disabled)
}

## -----------------------------------
## Method `RoogleSlides$enable()`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$enable()
}

## -----------------------------------
## Method `RoogleSlides$disable()`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$disable()
}

## -----------------------------------
## Method `RoogleSlides$withDocument(shareUrlOrDocId)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$withDocument(shareUrlOrDocId)
}

## -----------------------------------
## Method `RoogleSlides$findOrCreateDocument(title)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$findOrCreateDocument(title)
}

## -----------------------------------
## Method `RoogleSlides$findOrCloneTemplate(title, templateUri)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$findOrCloneTemplate(title, templateUri)
}

## -----------------------------------
## Method `RoogleSlides$tagsDefined()`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$tagsDefined()
}

## -----------------------------------
## Method `RoogleSlides$updateTaggedText(text, tagName)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$updateTaggedText(text, tagName)
}

## -----------------------------------
## Method `RoogleSlides$updateTaggedImage(absoluteFilePath, tagName, keepUpload)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$updateTaggedImage(absoluteFilePath, tagName, keepUpload)
}

## -----------------------------------
## Method `RoogleSlides$updateTaggedTable(longFormatTable, tagName, colWidths)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$updateTaggedTable(longFormatTable, tagName, colWidths)
}

## -----------------------------------
## Method `RoogleSlides$removeTags(confirm)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$removeTags(confirm)
}

## -----------------------------------
## Method `RoogleSlides$saveAsPdf(absoluteFilePath, uploadCopy)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$saveAsPdf(absoluteFilePath, uploadCopy)
}

## -----------------------------------
## Method `RoogleSlides$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$uploadSupplementaryFiles(absoluteFilePath, overwrite, duplicate)
}

## -----------------------------------
## Method `RoogleSlides$appendFormattedSlide(title, formattedTextDf, layout)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$appendFormattedSlide(title, formattedTextDf, layout)
}

## -----------------------------------
## Method `RoogleSlides$slideDimensions()`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$slideDimensions()
}

## -----------------------------------
## Method `RoogleSlides$slideLayouts()`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$slideLayouts()
}

## -----------------------------------
## Method `RoogleSlides$updateCitations(bibTexPath, citationStyle)`
## -----------------------------------
if (FALSE) {
# appropriate parameter values must be provided
instance$updateCitations(bibTexPath, citationStyle)
}