Use the default Maven repository location

It should be clear that running this function will affect files in the users filesystem.

Usage

.developer_mode()

Value

nothing. called for side effects

Details

The default Maven directory is located in user space and writing to it is forbidden by CRAN policies. This plugin is set up to use a cache directory for the local Maven repository but if you are developing Java code and using it in R then your Java build tools will be installing content to the default Maven directory, and not the CRAN sanctioned cache. Thus `rmaven` wont by default be able to pick up jar files locally installed through standard Java tooling. This function sets the `rmaven` repository location to the Maven standard location allowing local jar files to be used. Obviously this is not portable until the Java packages are deployed to a Maven repository and is only an useful option during development.