RoogleDocs

Programmatically substitute images, data and tables into a google doc.

Version: 0.1.0

Generated: 2022-05-19T09:47:25.609

Arguments

tokenDirectory
tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to normalizePath('~/.roogledocs', mustWork = FALSE)) - (java expects a String)

disabled
disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to getOption('roogledocs.disabled',FALSE)) - (java expects a boolean)

Details

R library to perform limited interactions with google docs (and maybe one day slides) in R via the Java API library. The purpose being to support google docs as a platform for interactive development and documentation of data analysis in R for scientific publication, although it is not limited to this purpose. The workflow supported is a parallel documentation and analysis where a team of people are working collaboratively on documentation, whilst at the same time analysis is being performed and results updated repeatedly as a result of new data. In this environment updating numeric results, tabular data and figures in word documents manually becomes annoying. With roogledocs you can automate this a bit like a RMarkdown document, but with the added benefit that the content can be updated independently of the analysis, by the wider team.

Methods

Constructors

J$RoogleDocs$new(tokenDirectory, disabled)

Static methods

none

Instance methods

instance$reauth()

instance$enable()

instance$disable()

instance$withDocument(shareUrlOrDocId)

instance$findOrCreateDocument(title)

instance$findOrCloneTemplate(title, templateUri)

instance$searchForDocuments(titleMatch)

instance$tagsDefined()

instance$updateTaggedText(text, tagName)

instance$updateTaggedImage(absoluteFilePath, tagName, dpi)

instance$revertTags()

instance$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)

instance$updateFigure(absoluteFilePath, figureIndex, dpi)

instance$saveAsPdf(absoluteFilePath)

instance$deleteDocument(docName, areYouSure)

instance$appendText(text, style)

instance$appendFormattedParagraph(formattedTextDf)

instance$clone()

instance$print()

Method new()

no description

Usage

J = roogledocs::JavaApi$get()
J$RoogleDocs$new(tokenDirectory, disabled)
        

Arguments

tokenDirectory the place to store authentication tokens. This should not be checked into version control. - (defaulting to normalizePath('~/.roogledocs', mustWork = FALSE)) - (java expects a String)

disabled a flag to switch roogledocs off (on a document by document basis, for testing or development. This can be set globally with `options('roogledocs.disabled'=TRUE)` - (defaulting to getOption('roogledocs.disabled',FALSE)) - (java expects a boolean)

Returns

R6 RoogleDocs object:

Examples

not available

Method reauth()

Reauthenticate the service deleting the existing OAuth tokens. Generally this would only be needed if application permission updates are needed in which case the directory can be manually deleted anyway, or if you want to switch google user without using a different tokenDirectory.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$reauth()
        

Arguments

none

Returns

R6 RoogleDocs object: a fluent method

Examples

not available

Method enable()

Enables roogledocs. It is likely one of `withDocument()`, `findOrCreateDocument()` or `findOrCloneTemplate()` methods will be needed to specify the document.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$enable()
        

Arguments

none

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method disable()

disables roogledocs temporarily only for this document. While disabled all calls to roogledocs will silently fail.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$disable()
        

Arguments

none

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method withDocument()

Select a document by its share url or id.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$withDocument(shareUrlOrDocId)
        

Arguments

shareUrlOrDocId the url from clicking a share button in google docs or an id from searchForDocuments() method - (java expects a String)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method findOrCreateDocument()

Search for or create a document if it is missing.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$findOrCreateDocument(title)
        

Arguments

title a document title. If there is an exact match in google drive then that document will be used otherwise a new one will be created. - (java expects a String)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method findOrCloneTemplate()

Search for or create a document as a copy of a template document if it is missing.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$findOrCloneTemplate(title, templateUri)
        

Arguments

title a document title. If there is an exact match in google drive then that document will be used otherwise a new one will be created. - (java expects a String)

templateUri the share link (or document id) of a template google document - (java expects a String)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method searchForDocuments()

Search for documents with the given title

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$searchForDocuments(titleMatch)
        

Arguments

titleMatch a string to be searched for as an approximate match. All results will be retrieved with document ids. - (java expects a String)

Returns

RDataframe: a dataframe containing id and name columns

Examples

not available

Method tagsDefined()

Finds tags defined in the current document

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$tagsDefined()
        

Arguments

none

Returns

RDataframe: a dataframe containing tag and count columns

Examples

not available

Method updateTaggedText()

Substitutes all occurrences of tag-name with the text parameter.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$updateTaggedText(text, tagName)
        

Arguments

text the value to replace the tag with (e.g. a result from analysis) (cannot be empty) - (java expects a String)

tagName the tag name - (java expects a String)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method updateTaggedImage()

Substitutes all occurrences of tag-name with an image from the local storage. There are limited circumstances in which using this is a good idea. It will almost always be better to use `updateFigure()` to insert an image by index. If you choose to ignore this warning, beware combining this with `updateFigure()` as potentially the figure indexes will change dynamically. You have been warned.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$updateTaggedImage(absoluteFilePath, tagName, dpi)
        

Arguments

absoluteFilePath a file path to an png image file. - (java expects a String)

tagName the tag name - (java expects a String)

dpi the dots per inch of the image in the document (defaults to 300) - (defaulting to 300) - (java expects a double)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method revertTags()

remove all tagged text and images inserted by roogledocs and returns the bare document. This does not affect figures and tables inserted by index (i.e. without tags) This is needed if content is being moved around as cut and paste of tagged content unfortunately removes the internal named range of the tag.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$revertTags()
        

Arguments

none

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method updateTable()

Update or insert a formatted table into the document. The table and formatting are described in a dataframe the format of which is documented in the as.long_format_table() method.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$updateTable(longFormatTable, tableIndex, colWidths, tableWidthInches)
        

Arguments

longFormatTable A dataframe consisting of the table content and formatting indexed by row and column. at a minimum this should have columns label,row,col, but may also include rowSpan,colSpan,fillColour, leftBorderWeight, rightBorderWeight, topBorderWeight, bottomBorderWeight, alignment (START,CENTER,END), valignment (TOP,MIDDLE,BOTTOM), fontName, fontFace, fontSize. - (java expects a RDataframe)

tableIndex what is the table index in the document? This can be left out for a new table at the end of the document. - (defaulting to -1) - (java expects a int)

colWidths A vector including the relative length of each column. This can be left out if longFormatTable comes from as.long_format_table - (defaulting to attr(longFormatTable,'colWidths')) - (java expects a RNumericVector)

tableWidthInches The final width of the table in inches (defaults to a size that fits in A4 page with margins) - (defaulting to 6.2) - (java expects a RNumeric)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method updateFigure()

Update or insert a figure in the document.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$updateFigure(absoluteFilePath, figureIndex, dpi)
        

Arguments

absoluteFilePath a file path to an png image file (only png is supported at this point). - (java expects a String)

figureIndex what is the figure index in the document? (This only counts inline images - and ignores absolutely positioned ones). leave out for a new image at the end of the document. - (defaulting to -1) - (java expects a int)

dpi the dots per inch of the image in the document (defaults to 300). the final size of the image in the doc will be determined by the image file dimensions and the dpi. - (defaulting to 300) - (java expects a double)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method saveAsPdf()

Saves a snapshot of the current google doc as a pdf to a local drive. This is mainly intended for testing.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$saveAsPdf(absoluteFilePath)
        

Arguments

absoluteFilePath - a file path to save the pdf. - (java expects a String)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method deleteDocument()

Deletes a google document by name.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$deleteDocument(docName, areYouSure)
        

Arguments

docName - the name of a document to delete. must be an exact and unique match. - (java expects a String)

areYouSure - a boolean check. - (defaulting to utils::askYesNo(paste0('Are you sure you want to delete ',docName),FALSE)) - (java expects a boolean)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method appendText()

Append text to the document with optional paragraph styling. If you run text blocks into each other without newlines the whole resulting paragraph will be styled. You would normally not want this so it is up to you to end paragraphs with a new line character, before changing styles.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$appendText(text, style)
        

Arguments

text - a single string with the text to append which may include newlines - (java expects a RCharacter)

style - one of NORMAL_TEXT, TITLE, SUBTITLE, HEADING_1, ... HEADING_6 - (defaulting to 'NORMAL_TEXT') - (java expects a RCharacter)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Method appendFormattedParagraph()

Append a new paragraph, with text from the 'label' column with optional formating in the other columns.

Usage

J = roogledocs::JavaApi$get()
instance = J$RoogleDocs$new(...);
instance$appendFormattedParagraph(formattedTextDf)
        

Arguments

formattedTextDf - a data frame containing the columns label, and optionally: link (as a URL), fontName, fontFace, fontSize. - (java expects a RDataframe)

Returns

R6 RoogleDocs object: itself - a fluent method

Examples

not available

Examples

if (FALSE) {
J = roogledocs::JavaApi$get();
instance = J$RoogleDocs$new(tokenDirectory, disabled)

## -----------------------------------
## Method `RoogleDocs$reauth`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$enable`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$disable`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$withDocument`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$findOrCreateDocument`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$findOrCloneTemplate`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$searchForDocuments`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$tagsDefined`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$updateTaggedText`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$updateTaggedImage`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$revertTags`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$updateTable`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$updateFigure`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$saveAsPdf`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$deleteDocument`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$appendText`
## -----------------------------------

## -----------------------------------
## Method `RoogleDocs$appendFormattedParagraph`
## -----------------------------------

}